first compiling version of plugin system
